pyproject.toml
[project]
name = "capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate==0.34.2",
    "bs4>=0.0.2",
    "datasets==3.0.1",
    "dotenv>=0.9.9",
    "evaluate==0.4.3",
    "flask>=3.1.2",
    "google-generativeai>=0.8.5",
    "huggingface-hub==0.26.2",
    "imbalanced-learn==0.12.3",
    "ipykernel>=7.1.0",
    "ipython>=9.7.0",
    "ipywidgets>=8.1.8",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.1",
    "langgraph>=1.0.1",
    "matplotlib==3.9.2",
    "nltk>=3.9.2",
    "numpy>=2.1.0",
    "openai>=2.7.1",
    "pandas==2.2.3",
    "scikit-learn==1.5.2",
    "torch==2.4.1",
    "torchvision==0.19.1",
    "transformers==4.45.2",
    "tweepy>=4.16.0",
]
